requirements.txt
absl-py==1.2.0
aniso8601==9.0.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
flatbuffers==1.12
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pip==22.1.2
Pillow
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==63.4.1
six==1.16.0
tensorflow-cpu
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
